requirements.txt.txt
Flask==3.0.3
Werkzeug==3.0.3
mysql-connector-python==9.0.0
razorpay==1.4.2
requests==2.32.3
gunicorn==23.0.0
